backend/pyproject.toml
[tool.poetry]
name = "xray-panel"
version = "0.1.0"
description = "Xray Control Plane - Multi-node proxy management panel"
authors = ["Panel Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
redis = "^5.0.1"
celery = "^5.3.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
httpx = "^0.26.0"
aiofiles = "^23.2.1"
pyyaml = "^6.0.1"
jinja2 = "^3.1.3"
qrcode = {extras = ["pil"], version = "^7.4.2"}
python-telegram-bot = "^20.8"
pyotp = "^2.9.0"
cryptography = "^42.0.0"
grpcio = "^1.60.0"
grpcio-tools = "^1.60.0"
protobuf = "^4.25.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
black = "^24.1.1"
ruff = "^0.1.14"
mypy = "^1.8.0"

[tool.poetry.scripts]
panel = "app.cli.main:app"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
